Outputstream: Fix endianness issues when writing array lengths
When writing a 2 or 4 byte array length (configured by arrayLengthWidth),
the 16 or 32bit unsigned integer is always written as / converted
to (depending on the endianness of the machine) big endian.

According to the someip spec, the payload's endianness is configurable,
and in the case of little endian, the serializer wrongly assumes
big endian as the payloads endianness.

This causes the deserializer to read the bytes specifying the length of
an array in the wrong byte order, if the payload is little endian,
leading to an error on the consuming side (code is REMOTE_ERROR and not
easy to track).