Revert setting the default width of a serialized enumeration
Setting the width as this code did narrows the enumeration value when it
isn't set to the default and a deployment object is unused.

e.g. With that static_cast, even the enumeration is a uint32_t for
example, it would be narrowed to a uint8_t and serialization would fail.